+++ Login-Permission-Project-frontend/src/assets/styles/main.css
@@ -10,3 +10,195 @@
     background-color:#faaa65; /*#006284;#1E88A8*/
     border-color: #faaa65; /*#006284;#1E88A8*/  /*#91B493*/
 }
+
+.container {
+    padding: 20px;
+}
+
+.content-wrapper {
+    background-color: #ffffff;
+    border-radius: 8px;
+    padding: 1.5rem;
+    box-shadow: 0 2px 4px rgba(0, 0, 0, 0.05);
+}
+
+.nav-tabs .nav-link {
+    color: #334255;
+    border: none;
+    padding: 0.75rem 1.5rem;
+    font-size: 1.1rem;
+    cursor: pointer;
+}
+
+.nav-tabs .nav-link.active {
+    color: #334255;
+    font-weight: 600;
+    border-bottom: 2px solid #334255;
+    background: none;
+}
+
+.add-employee-btn,
+.add-role-btn {
+    color: #334255;
+    font-weight: 500;
+    padding: 0.5rem 1rem;
+    border: 2px solid #334255;
+    border-radius: 6px;
+    transition: all 0.3s ease;
+}
+
+.add-employee-btn:hover,
+.add-role-btn:hover {
+    background-color: #334255;
+    color: #ffffff;
+}
+
+.search-btn {
+    background-color: #334255;
+    border-color: #334255;
+    padding: 0.75rem;
+    height: 58px;
+    width: 58px;
+}
+
+.search-btn:hover {
+    background-color: #FFCD50;
+    border-color: #FFCD50;
+    color: #334255;
+}
+
+.form-floating>.form-select,
+.form-floating>.form-control {
+    padding-top: 1.625rem;
+    padding-bottom: 0.625rem;
+    height: 100%;
+}
+
+.form-floating label {
+    color: #334255;
+    font-weight: 500;
+}
+
+.table {
+    margin-bottom: 0;
+    white-space: nowrap;
+}
+
+.table th {
+    background-color: #f8f9fa;
+    font-weight: 600;
+    padding: 1rem;
+    border-bottom: 2px solid #dee2e6;
+}
+
+.table td {
+    padding: 1rem;
+    vertical-align: middle;
+}
+
+.btn-link {
+    color: #334255;
+    padding: 0.375rem;
+    transition: all 0.2s ease;
+}
+
+.btn-link:hover {
+    color: #FFCD50;
+    transform: scale(1.1);
+}
+
+.btn-link.text-danger:hover {
+    color: #dc3545;
+}
+
+.badge {
+    font-size: 0.8rem;
+    padding: 0.5em 0.7em;
+    margin: 0.2em;
+    background-color: #6c757d;
+}
+
+@media (max-width: 768px) {
+    .container {
+        padding: 1rem;
+    }
+
+    .content-wrapper {
+        padding: 1rem;
+    }
+
+    .form-floating {
+        margin-bottom: 1rem;
+    }
+
+    .add-employee-btn,
+    .add-role-btn {
+        width: 100%;
+        margin-bottom: 1rem;
+    }
+
+    .table th,
+    .table td {
+        font-size: 0.9rem;
+        padding: 0.5rem;
+    }
+}
+
+/* Modal 樣式 */
+.modal-content {
+    border-radius: 8px;
+    border: none;
+}
+
+.modal-header {
+    background-color: #f8f9fa;
+    border-bottom: 2px solid #FFCD50;
+}
+
+.modal-header .modal-title {
+    color: #334255;
+    font-weight: 600;
+}
+
+.modal-footer {
+    border-top: 1px solid #dee2e6;
+    padding: 1rem;
+}
+
+.permission-checkboxes {
+    max-height: 200px;
+    overflow-y: auto;
+    padding: 0.5rem;
+    border: 1px solid #dee2e6;
+    border-radius: 4px;
+}
+
+.form-check {
+    margin-bottom: 0.5rem;
+}
+
+.form-check-input:checked {
+    background-color: #334255;
+    border-color: #334255;
+}
+
+.btn-primary {
+    background-color: #334255;
+    border-color: #334255;
+}
+
+.btn-primary:hover {
+    background-color: #FFCD50;
+    border-color: #FFCD50;
+    color: #334255;
+}
+
+.btn-danger {
+    background-color: #dc3545;
+    border-color: #dc3545;
+}
+
+.btn-danger:hover {
+    background-color: #bb2d3b;
+    border-color: #b02a37;
+}
