+++ Login-Permission-Project-frontend/src/components/HumanResourcesPage/LoginRecord/LoginRecordComponet.vue
