+++ Login-Permission-Project-frontend/src/components/ManagementPage/Role/abc.vue
@@ -1,0 +1,5 @@
+<template>
+    <div>
+        <p>abc</p>
+    </div>
+</template>
