+++ Login-Permission-Project-frontend/src/components/auth/Register.vue
